+++ libpsx-rs/src/system/gpu/controllers.rs
@@ -34,7 +34,7 @@ pub(crate) fn run(context: &ControllerContext, event: Event) -> ControllerResult
     }
 }
 
-fn run_time(state: &State, video_backend: &VideoBackend, duration: f64) -> ControllerResult<()> {
+fn run_time(state: &State, video_backend: &VideoBackend, duration: f32) -> ControllerResult<()> {
     let controller_state = &mut state.gpu.controller_state.lock();
     controller_state.clock += duration;
